requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.5
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-filter==25.1
django-js-asset==3.1.2
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.20.1
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
